Add unit tests for the GoToSite updates
This test is modified from the unit test `test_get` in the same class. 

This unit tests verifies that a url with no path, a url with a subpath, a url with a query, and a url with a query and a subpath are all correctly redirected when used with @@GoToSite.